fix(ci): python-full-coverage — relocate overflow test to keep part_054 under 1500-line guardrail
The review-fix regression test appended to test_worker_part_054.py pushed it
from 1493 to 1501 lines, tripping
test_first_party_code_files_stay_under_line_limit (MAX_FIRST_PARTY_FILE_LINES
= 1500). Move test_monitor_recovery_handoff_failure_error_code_rejects_missing_reason_code
into the sibling test_worker_part_055.py (607 lines, monitor-recovery themed)
so the same behavior is still asserted and part_054 drops back to 1493 lines.
